+++ lib/dead_end/display_invalid_blocks.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "banner"
 require_relative "capture_code_context"
 require_relative "display_code_with_line_numbers"
 
@@ -54,43 +55,7 @@ module DeadEnd
     end
 
     def banner
-      case @invalid_obj.error_symbol
-      when :missing_end
-        <<~EOM
-          DeadEnd: Missing `end` detected
-
-          This code has a missing `end`. Ensure that all
-          syntax keywords (`def`, `do`, etc.) have a matching `end`.
-        EOM
-      when :unmatched_syntax
-        case @invalid_obj.unmatched_symbol
-        when :end
-          <<~EOM
-            DeadEnd: Unmatched `end` detected
-
-            This code has an unmatched `end`. Ensure that all `end` lines
-            in your code have a matching syntax keyword  (`def`,  `do`, etc.)
-            and that you don't have any extra `end` lines.
-          EOM
-        when :|
-          <<~EOM
-            DeadEnd: Unmatched `|` character detected
-
-            Example:
-
-              `do |x` should be `do |x|`
-          EOM
-        when :"}"
-          <<~EOM
-            DeadEnd: Unmatched `}` character detected
-
-            This code has an unmatched `}`. Ensure that opening curly braces are
-            closed: `{ }`.
-          EOM
-        else
-          "DeadEnd: Unmatched `#{@invalid_obj.unmatched_symbol}` detected"
-        end
-      end
+      Banner.new(invalid_obj: @invalid_obj).call
     end
 
     def indent(string, with: "    ")