+++ lib/dead_end/who_dis_syntax_error.rb
@@ -8,6 +8,7 @@ module DeadEnd
   #   puts WhoDisSyntaxError.new("def foo;").call.error_symbol
   #   # => :missing_end
   class WhoDisSyntaxError < Ripper
+    CHARACTERS = {"{": :"}", "}": :"{", "[": :"]", "]": :"[", "(": :")", ")": :"("}
     class Null
       def error_symbol
         :missing_end
@@ -59,6 +60,12 @@ module DeadEnd
         @unmatched_symbol = :end
         @error_symbol = :unmatched_syntax
       when /unexpected .* expecting ['`]?(?<unmatched_symbol>[^']*)/
+        if $1
+          character = $1.to_sym
+          @unmatched_symbol = CHARACTERS[character] || character
+        end
+        @error_symbol = :unmatched_syntax
+      when /unexpected '(?<unmatched_symbol>.*)'/
         @unmatched_symbol = $1.to_sym if $1
         @error_symbol = :unmatched_syntax
       when /unexpected `end'/, # Ruby 2.7 and 3.0